+++ packages/dartblaze/lib/dartblaze.dart
@@ -3,7 +3,8 @@
 /// More dartdocs go here.
 library;
 
-/// Firebase
+/// Core
+export 'src/environment.dart';
 export "src/firebase.dart";
 
 /// Firestore
@@ -15,3 +16,5 @@ export "src/firestore/firestore_path_parser.dart";
 /// HTTP
 export "src/http/annotations.dart";
 export 'src/utils/request_body_accessor.dart';
+export 'src/http/firebase_auth_validator.dart';
+export 'package:openid_client/src/model.dart' show IdToken;