+++ packages/dartblaze/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+- Firebase Auth integration for Http functions
+- Automatic detection of project id
+
 ## 0.0.1
 
-- Initial version.
+- Http functions
+- Firestore triggers
+